PCLOUDS-1602: SFM refactor - fix imports
